Support ignoring file paths
One or more `-i` options can be provided, which are taken as entries
for a `gitignore` file that is used to filter rendered entries
from the diff.

For example `-i com/foo/F.class.* -i com/foo/impl/`  would filter
all renderings of the class `com.foo.F` and of the contents of
the `impl` package. `-i *.class.scalap` would filter the diff
of the Scala Signature.